refactor: move config to common
