+++ apps/worker/src/asset_processor/asset.processor.module.ts
@@ -5,11 +5,11 @@ https://docs.nestjs.com/modules
 import { BullModule } from '@nestjs/bullmq';
 import { Module } from '@nestjs/common';
 import { RedisModule } from '@liaoliaots/nestjs-redis';
-import { ConfigModule } from '../../../api/src/config/config.module';
-import { ConfigService } from '../../../api/src/config/config.service';
 import { QueueConstants } from '../../../../libs/common/src';
 import { AssetProcessorService } from './asset.processor.service';
 import { IpfsService } from '../../../../libs/common/src/utils/ipfs.client';
+import { ConfigModule } from '../../../../libs/common/src/config/config.module';
+import { ConfigService } from '../../../../libs/common/src/config/config.service';
 
 @Module({
   imports: [
