+++ apps/worker/src/batch_announcer/batch.announcer.ts
@@ -7,7 +7,7 @@ import Redis from 'ioredis';
 import { PalletSchemasSchema } from '@polkadot/types/lookup';
 import { hexToString } from '@polkadot/util';
 import { BlockchainService } from '../blockchain/blockchain.service';
-import { ConfigService } from '../../../api/src/config/config.service';
+import { ConfigService } from '../../../../libs/common/src/config/config.service';
 import { IBatchAnnouncerJobData } from '../interfaces/batch-announcer.job.interface';
 import { IPublisherJob } from '../interfaces/publisher-job.interface';
 import { IpfsService } from '../../../../libs/common/src/utils/ipfs.client';
