+++ apps/worker/src/monitor/status.monitor.module.ts
@@ -7,8 +7,8 @@ import { Module } from '@nestjs/common';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { RedisModule } from '@liaoliaots/nestjs-redis';
 import { StatusMonitoringService } from './status.monitor.service';
-import { ConfigModule } from '../../../api/src/config/config.module';
-import { ConfigService } from '../../../api/src/config/config.service';
+import { ConfigModule } from '../../../../libs/common/src/config/config.module';
+import { ConfigService } from '../../../../libs/common/src/config/config.service';
 import { BlockchainModule } from '../blockchain/blockchain.module';
 import { QueueConstants } from '../../../../libs/common/src';
