+++ libs/common/src/interfaces/capacity-limit.interface.ts
