+++ libs/common/src/utils/ipfs.client.ts
@@ -9,7 +9,7 @@ import { blake2b256 as hasher } from '@multiformats/blake2/blake2b';
 import { create } from 'multiformats/hashes/digest';
 import { randomUUID } from 'crypto';
 import { base58btc } from 'multiformats/bases/base58';
-import { ConfigService } from '../../../../apps/api/src/config/config.service';
+import { ConfigService } from '../config/config.service';
 
 export interface FilePin {
   cid: string;
